requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
distro==1.9.0
Django==5.1.3
djangorestframework==3.15.2
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
pydantic==2.10.2
pydantic_core==2.27.1
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.2
typing_extensions==4.12.2
httpx
PyJWT
pymongo
python-dateutil

